Overhaul Lang handling
- Replace enum values with direct yaml checks (improves performance on
startup, by reducing the number of times the yaml and plugin are loaded)
- Hardcode colour codes into the yaml
- Load and store a small number of reflection-generated player chat
packets on startup (reduce performance load of reflection — I did some
benchmarking on this and reflection took more than 4000 times longer
:skull: :skull:
- Set up Action Bar messaging! Return messages for world interactions
(i.e. punching the easel) are now sent via a chat packet to the
player’s action bar. This is easier to see for the player, but also
less spammy and won’t fill up their chat. More messages may be moved
the action bar in future.